// frontend/src/api-generated/models/nhatKyThaoTacHeThongResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface NhatKyThaoTacHeThongResponse {
  id?: number;
  idTaiKhoanThaoTac?: number;
  phuongThucApi?: string;
  endpointApi?: string;
  thucTheTacDong?: string;
  idBanGhi?: number;
  duLieuTruoc?: string;
  duLieuSau?: string;
  diaChiIp?: string;
  thoiGianThaoTac?: string;
}